Refactor qualifier into subpackage, remove component subpackage
