+++ clinlp/__init__.py
@@ -1,3 +1,4 @@
-import clinlp.component
-
 from .language import Clinlp
+from .normalizer import Normalizer
+from .qualifier import ContextMatcher
+from .sentencizer import Sentencizer
