+++ clinlp/qualifier/__init__.py
@@ -1,13 +1,11 @@
 from spacy.tokens import Span
 
-from .normalizer import Normalizer
-from .qualifier import (
+from .qualifier import Qualifier, QualifierDetector
+from .rule_based import (
     QUALIFIERS_ATTR,
     ContextMatcher,
     ContextRule,
     ContextRuleDirection,
-    Qualifier,
 )
-from .sentencizer import Sentencizer
 
 Span.set_extension(name=QUALIFIERS_ATTR, default=None)
